### 2023-04-26 #5413 @editorialbot
[PRE REVIEW]: OmniTrax: A deep learning-driven multi-animal tracking and pose-estimation add-on for Blender
**Submitting author:** @FabianPlum (<a href="http://orcid.org/0000-0003-1012-6646">Fabian Plum</a>)
**Repository:** https://github.com/FabianPlum/OmniTrax
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.2
