+++ paper/paper.bib
@@ -45,6 +45,7 @@
     title = {Real-time, low-latency closed-loop feedback using markerless posture tracking},
     journal = {eLife},
     year = {2020},
+    url = {https://doi.org/10.7554/eLife.61909},
 }
 
 @article{Plum2021,
