+++ paper/paper.bib
@@ -37,7 +37,7 @@
     author = {Alexander Mathis and Pranav Mamidanna and Kevin M. Cury and Taiga Abe and Venkatesh N. Murthy and Mackenzie W. Mathis and Matthias Bethge},
     journal = {Nature Neuroscience},
     year = {2018},
-    url = {https://www.nature.com/articles/s41593-018-0209-y}
+    doi = {10.1038/s41593-018-0209-y}
 }
 
 @article{Kane2020dlclive,
@@ -45,7 +45,7 @@
     title = {Real-time, low-latency closed-loop feedback using markerless posture tracking},
     journal = {eLife},
     year = {2020},
-    url = {https://doi.org/10.7554/eLife.61909},
+    doi = {10.7554/eLife.61909},
 }
 
 @article{Plum2021,
