### 2023-05-30 #5413 @Kevin-Mattheus-Moerman
@gineshidalgo99 @davidflam @jgraving @py-motmetrics @SwathiSheshadri @timohueser @MMathisLab @jeylau @jonmatthis @aaroncherian would you be interested in reviewing the following submission for [JOSS (The Journal of Open Source Software)](https://joss.theoj.org/about): "OmniTrax: A deep learning-driven multi-animal tracking and pose-estimation add-on for Blender"?

JOSS reviews focus on evaluating [the software](https://github.com/FabianPlum/OmniTrax) and reviewing [a short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05413/joss.05413/10.21105.joss.05413.pdf). Reviews typically take around 3 weeks, but we can be more flexible if you need more time. You can let me know here if you are interested. Thanks!

FYI, here is an example of a review issue (this here is a pre-review issue where we find/assign reviewers): https://github.com/openjournals/joss-reviews/issues/5278, as you can see reviews are guided by a checklist that each reviewer completes. More on our review process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
