### 2023-06-12 #5413 @jonmatthis
I can help :)

Btw - Aaron Cherian is my graduate student, so it probably wouldn't make
sense to have both of us on. He'll certainly have a hand in my review

On Fri, May 19, 2023 at 10:52 AM Kevin Mattheus Moerman <
***@***.***> wrote:

> @SwathiSheshadri <https://github.com/SwathiSheshadri> @timohueser
> <https://github.com/timohueser> @MMathisLab
> <https://github.com/MMathisLab> @jeylau <https://github.com/jeylau>
> @jonmatthis <https://github.com/jonmatthis> @aaroncherian
> <https://github.com/aaroncherian> would you be interested in reviewing
> the following submission for JOSS (The Journal of Open Source Software)
> <https://joss.theoj.org/about>: "OmniTrax: A deep learning-driven
> multi-animal tracking and pose-estimation add-on for Blender"?
>
> JOSS reviews focus on evaluating the software and reviewing a short paper.
> Reviews typically take around 3 weeks, but we can be more flexible if you
> need more time. You can let me know here if you are interested. Thanks!
>
> FYI, here is an example of a review issue (this here is a pre-review issue
> where we find/assign reviewers): #5278
> <https://github.com/openjournals/joss-reviews/issues/5278>, as you can
> see reviews are guided by a checklist that each reviewer completes. More on
> our review process here:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
