### 2023-06-13 #5549 @editorialbot
[REVIEW]: OmniTrax: A deep learning-driven multi-animal 1 tracking and pose-estimation add-on for Blender
**Submitting author:** @FabianPlum (<a href="http://orcid.org/0000-0003-1012-6646">Fabian Plum</a>)
**Repository:** https://github.com/FabianPlum/OmniTrax
**Branch with paper.md** (empty if default branch):
**Version:** V_1.0.0
