### 2023-07-02 710e830
Fixed issue where masks are no longer optional -resolves #4
