### 2023-07-02 84f206b
Updated path handling for yolo .names and .data files -resolves #5
