### 2023-07-02 #5549 @FabianPlum
Hi all! Thanks for your help on the review!

Just FYI, I recently fixed a few minor bugs relating to masking and path handling issues a user kindly reported.
I now added an updated release that includes these fixes:
[OmniTrax Version 0.2.3](https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.2.3)

All the best
Fabi
