### 2023-08-16 #5549 @lucasmiranda42
Dear all,

apologies for the delay. I finish reading the paper and commenting checking what I could, but so far I didn't manage to install the package due to OS incompatibilities (I didn't have access to a Windows machine, and I was aiming to test it using a Parallels VM in MacOS, which unfortunately did not work). I just borrowed a Windows laptop that has a GPU, and I intend to test installation on that machine. I'll keep you posted!

Best,
Lucas
