### 2023-09-04 FabianPlum/OmniTrax#6 @lucasmiranda42
[JOSS] Missing contributing guidelines
Hey there! One of your [JOSS](https://github.com/openjournals/joss-reviews/issues/5549) reviewers here. While the readme file is solid, the package seems to be missing clear contribution guidelines for people who would like to send pull requests.

I would strongly recommend adding one (you can get a pretty good idea on what to include [here](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)).

Best!
Lucas
