### 2023-09-04 #5549 @lucasmiranda42
Dear @FabianPlum, dear all,

After testing on several machines I can confirm I managed to successfully install the package! Issues will shortly follow :)

Best, and apologies for the massive delays,
Lucas
