### 2023-09-04 FabianPlum/OmniTrax#7 @lucasmiranda42
[JOSS] OS compatibility
Lucas, from your [JOSS](https://github.com/openjournals/joss-reviews/issues/5549) submission, again.

The package so far looks awesome, to be honest, but it took me a while to get my hands on a proper machine to install it. What are your thoughts / what would the main issues be with making at least the CPU version compatible with other operating systems (ie MacOS, Linux)?

Best!
