### 2023-09-05 c38181f
Created contributing guideline | -resolves #6
