### 2023-09-05 #5549 @FabianPlum
Thanks so much for the kind feedback @lucasmiranda42!
I added a short contributing guide and will have a look if I can make a basic CPU version of OmniTrax run on Ubuntu. As I don't have access to any MacOS systems, I can't really promise a working and tested implementation on that side.

All the best
Fabian
