### 2023-10-09 #5549 @lucasmiranda42
Dear @Kevin-Mattheus-Moerman, dear all,

thank you for your comment. My only concern at the moment is the open issue regarding OS compatibility. If this is too hard to accomplish, a justification why would also suffice on my end. Other than that (which I encompassed under the functionality box) this has green light on my end :)

Best!
Lucas
