### 2023-10-09 #5549 @FabianPlum
Hi @lucasmiranda42,

Thanks so much for your feedback! I'm in the process of writing up my PhD thesis so I'm a little slow to implement new functionality at the moment. One issue with getting this to run quickly under Linux lies in the way the neural network side to run YOLO is easier to compile for windows where I built the basic libraries that seem to run well on all tested systems. On the general Linux side I'd have to spend more time reading into how this could be built on the fly (or solved differently altogether) as the system landscape is much more heterogeneous.

I will look into this again the second half of this week!

Thanks for all your time and help with the review.

All the best
Fabi
