### 2023-10-11 4dad503
Added (CPU) support for Ubuntu systems | -closes #7
