### 2023-10-11 FabianPlum/OmniTrax#7 @FabianPlum
Hi @lucasmiranda42,

Thanks a lot for all your time and feedback. I finally got around to putting together a working version of OmniTrax for Linux systems (tested on Ubuntu 18.04).

For now the installation defaults to CPU only mode and has been tested and verified with [Blender 2.92](https://www.blender.org/download/releases/2-92/).
While all functionality of the Windows version is now given under Ubuntu (18.04) as well, CPU inference for tracking is extremely slow (~0.2 fps on a CPU vs ~20 to 40 fps on a GPU under Windows 10/11), while pose inference via tensorflow and dlc-live on a CPU runs perfectly well.

You can however relatively easily re-make darknet with GPU, OPENMP, AVX, OPENCV and whatever else you desire in its respective folder (see Makefile in OmniTrax/darknet **AFTER** triggering the installation via **check_packages.py** or by enabling the addon in Blender, as when installed on Linux, OmniTrax will clone the latest darknet repo, modify the Makefile and make darknet on the fly), given your system environment is set up accordingly. The omnitrax base installation takes care of everything to get things up and running - tweaking performance is then up to the end-user where desired.

I hope this implementation is satisfactory and enables further users to make use of this toolset!

All the best
Fabi
