### 2023-10-25 68ed39f
Updated references in paper
