### 2023-10-27 #5549 @jonmatthis
Got it running on Blender, but there's a filepath error when I try to run it:

![image](https://github.com/openjournals/joss-reviews/assets/15314521/1a775e55-cc46-472e-b27e-a9d1c13b4325)

![image](https://github.com/openjournals/joss-reviews/assets/15314521/efdab017-f2df-4171-a844-ce9b9beb7af7)

![image](https://github.com/openjournals/joss-reviews/assets/15314521/12aafb09-763e-4d0a-8523-8fb0078a743b)

Error:

```
Running inference on:  [LogicalDevice(name='/device:CPU:0', device_type='CPU'), LogicalDevice(name='/device:GPU:0', device_type='GPU')]

INFO: Initialised darknet network found!

Traceback (most recent call last):
  File "C:\Users\jonma\AppData\Roaming\Blender Foundation\Blender\3.3\scripts\addons\omni_trax\__init__.py", line 132, in execute
    video = bpy.path.abspath(clip.filepath)
AttributeError: 'NoneType' object has no attribute 'filepath'
Error: Python: Traceback (most recent call last):
  File "C:\Users\jonma\AppData\Roaming\Blender Foundation\Blender\3.3\scripts\addons\omni_trax\__init__.py", line 132, in execute
    video = bpy.path.abspath(clip.filepath)
AttributeError: 'NoneType' object has no attribute 'filepath'
```
