### 2023-10-27 #5549 @FabianPlum
Hi Jon! Could you send me the video you are running the tracking on?
I can only reproduce the error when I have not loaded any video
