### 2023-10-27 383cb92
Updated references to network file-requirements
