### 2023-10-27 #5549 @jonmatthis
Blender crash on run -

Please specify which combination of video and cfg etc settings is expected to work properly
