### 2023-10-27 #5549 @FabianPlum
Your configuration looks good, but in the screenshots above it appears you haven't opened any video in your project.
With your configuration above, everything should work as expected with any of the example videos provided, e.g: [example_ant_recording.mp4](https://github.com/FabianPlum/OmniTrax/blob/main/images/example_ant_recording.mp4)
