### 2023-10-27 #5549 @jonmatthis
Cool thanks -

In general, your intro tutorial should not have *choices* in it

You have a few spots where you're like "Here's a bunch of videos/config files - pick one and put it in" - that's not great because then A - You don't know what the user is going to do and B - the user doesn't know what to do!

The consequence of that confusion is the conversation we just had.

You should update your tutorial to specify *exactly* which files should be input and then *exactly* what the expected output should be.
