### 2023-10-27 #5549 @jonmatthis
Yeah, the implicit 'agreement' of an intro tutorial is "If you do exactly these steps, it should work"

It makes it much cognitively easier for the user (I don't have enough context to make decisions yet!) and MUCH easier for you to field complaints (the vast majority of which will be of the form "I tried it and it didn't work," bless em)

when your intro tutorial is locked in, it minimizes the number of ways I can screw it up , which makes it harder for you to tell if the error is because your code is broken vs a standard PEBKAC error

If you haven't seen it, I highly recommend checking out https://diataxis.fr for a great framework for making good docs/tutorials etc
