### 2023-10-27 #5549 @FabianPlum
Thanks for the resource!
I'll update the intro tutorial within the main readme to be more straightforward and the extended versions to be less vague
