### 2023-10-27 47526f5
Simplified "test drive" tutorial
