### 2023-10-27 cf2e2e9
Updated extended pose-estimation tutorial
