### 2023-10-27 15d1c90
Updated references to used video and trained network
