### 2023-10-27 #5549 @FabianPlum
@jonmatthis All docs updated!
Should be less convoluted and more fail-safe-ish now
