+++ paper/paper.bib
@@ -61,13 +61,13 @@
     journal = {{PeerJ}}
 }
 
-@misc{Plumetal2023,
-    author = {Fabian Plum and Rene Bulla and Hendrik Beck and Natalie Imirzian and David Labonte},
-    title = {replicAnt: A pipeline for generating annotated images of animals in complex environments using Unreal Engine},
-    elocation-id = {2023.04.20.537685},
-    year = {2023},
-    doi = {10.1101/2023.04.20.537685},
-    publisher = {Cold Spring Harbor Laboratory},
-    URL = {https://www.biorxiv.org/content/10.1101/2023.04.20.537685v2},
-    journal = {bioRxiv}
+@article{Plumetal2023,
+    title = {replicAnt: a pipeline for generating annotated images of animals in complex environments using Unreal Engine},
+    author = {Plum, Fabian and Bulla, René and Beck, Hendrik K and Imirzian, Natalie and Labonte, David},
+    doi = {10.1038/s41467-023-42898-9},
+    issn = {2041-1723},
+    journal = {Nature Communications},
+    url = {https://doi.org/10.1038/s41467-023-42898-9},
+    volume = {14},
+    year = {2023}
     }
