### 2023-12-04 550a2ee
Added CUDA version checker to ensure the system environment is set up correctly
