### 2023-12-04 40ee98c
Updated installation and package checking routine to include CUDA checks
