### 2023-12-04 cdfcfec
Added improved exception handling for setup_state file
