### 2023-12-04 15efec5
Locked off CPU compute version when using blender version other than 2.92.0 and added improved error handling
