### 2023-12-05 0cd2fe0
Updated exception handling for loading YOLO models
