### 2023-12-05 c41f76b
Improved exception handling for YOLO and DLC use
