+++ __init__.py
@@ -106,7 +106,7 @@ class OMNITRAX_OT_DetectionOperator(bpy.types.Operator):
         if not os.path.isfile(yolo_cfg_path_temp) or yolo_cfg_path_temp.split(".")[-1] != "cfg":
             self.report({'ERROR'}, 'Provide the .cfg file of a suitable YOLOv3 or V4 model')
             return {'CANCELLED'}
-        
+
         yolo_weights_path_temp = bpy.path.abspath(context.scene.detection_weights_path)
         if not os.path.isfile(yolo_weights_path_temp) or yolo_weights_path_temp.split(".")[-1] != "weights":
             self.report({'ERROR'}, 'Provide the .weights file of a trained YOLOv3 or V4 model')
@@ -331,10 +331,15 @@ class OMNITRAX_OT_DetectionOperator(bpy.types.Operator):
             raise ValueError("Invalid data file path `" +
                              os.path.abspath(yolo_paths.data) + "`")
         if network is None:
-            network, class_names, _ = darknet.load_network(yolo_paths.cfg,
-                                                           yolo_paths.data,
-                                                           yolo_paths.weights,
-                                                           batch_size=1)
+            try:
+                network, class_names, _ = darknet.load_network(yolo_paths.cfg,
+                                                               yolo_paths.data,
+                                                               yolo_paths.weights,
+                                                               batch_size=1)
+            except Exception as e:  # work on python 3.x
+                self.report({'ERROR'}, e)
+                return {'CANCELLED'}
+
         if altNames is None:
             try:
                 with open(yolo_paths.data) as metaFH:
@@ -353,8 +358,9 @@ class OMNITRAX_OT_DetectionOperator(bpy.types.Operator):
                                 altNames = [x.strip() for x in namesList]
                     except TypeError:
                         pass
-            except Exception:
-                pass
+            except Exception as e:  # work on python 3.x
+                self.report({'ERROR'}, e)
+                return {'CANCELLED'}
 
         print("Starting the YOLO loop...")
 
@@ -383,176 +389,181 @@ class OMNITRAX_OT_DetectionOperator(bpy.types.Operator):
         executed_from_frame = bpy.context.scene.frame_current
 
         while True:
-            if tracker_continue:
-                if frame_counter == context.scene.frame_end + 3 - executed_from_frame:
+            try:
+                if tracker_continue:
+                    if frame_counter == context.scene.frame_end + 3 - executed_from_frame:
+                        bpy.context.scene.frame_current = context.scene.frame_end
+                        break
+                elif frame_counter == context.scene.frame_end + 3 - context.scene.frame_start:
                     bpy.context.scene.frame_current = context.scene.frame_end
                     break
-            elif frame_counter == context.scene.frame_end + 3 - context.scene.frame_start:
-                bpy.context.scene.frame_current = context.scene.frame_end
-                break
 
-            prev_time = time.time()
-            ret, frame_read = cap.read()
-            if not ret:
-                break
+                prev_time = time.time()
+                ret, frame_read = cap.read()
+                if not ret:
+                    break
 
-            clip_width = frame_read.shape[1]
-            clip_height = frame_read.shape[0]
-            frame_rgb = cv2.cvtColor(frame_read, cv2.COLOR_BGR2RGB)
+                clip_width = frame_read.shape[1]
+                clip_height = frame_read.shape[0]
+                frame_rgb = cv2.cvtColor(frame_read, cv2.COLOR_BGR2RGB)
 
-            # apply mask, if defined
-            if img_mask is not None:
-                frame_rgb = cv2.bitwise_and(frame_rgb, frame_rgb, mask=img_mask)
+                # apply mask, if defined
+                if img_mask is not None:
[…]
