### 2023-12-05 0b7d3dc
Updated setup_state.txt generation
