### 2023-12-05 #5549 @FabianPlum
Hi @Kevin-Mattheus-Moerman and @jonmatthis,
I updated the OmniTrax documentation, installation, and improved the overall exception handling / error prompting throughout key pieces of functionality. I made a new release with these updates so the full installation can be tested.

[OmniTrax V_0.3.1](https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.3.1)

To quickly recap - OmniTrax GPU runs best within Blender 3.3 and 3.2 to ensure both tracking and pose estimation run smoothly. CPU inference is now disbaled in 3.3 by default as it does more harm than good to keep that option. For a generally stable version and both CPU and GPU inference options, please use Blender version 2.92.0, as stated in the documentation.

Please let me know if this addresses your concerns and runs as expected.

Thanks for all your help in making this repo as robust and user-friendly as possible.

All the best,
Fabi
