### 2023-12-05 5a75e65
Added more functionality to export data loader and visualisation
