### 2024-01-08 #5549 @jonmatthis
Got it working!

I finally had a chance to sit down and sort out the CUDA version graveyard on my PC and after that it was relatively clear sailing

review incoming
