### 2024-02-12 #5549 @Kevin-Mattheus-Moerman
@cellistigs, @edeno @gineshidalgo99 @davidflam @jgraving @py-motmetrics @SwathiSheshadri @timohueser @jeylau @aaroncherian @rizarae-p @maxsitt @animikhaich @aaa34169 @sfmig would you be interested in reviewing the following submission for [JOSS (The Journal of Open Source Software)](https://joss.theoj.org/about): "OmniTrax: A deep learning-driven multi-animal tracking and pose-estimation add-on for Blender"?

JOSS reviews focus on evaluating [the software](https://github.com/FabianPlum/OmniTrax) and reviewing [a short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05549/joss.05549/10.21105.joss.05549.pdf). Reviews typically take around 3 weeks, but we can be more flexible if you need more time. You can let me know here if you are interested. Thanks!

As you can see here, this review is already ongoing. One reviewer has completed there review already, however we are still in need of an additional reviewer.

More on our review process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

If you are interested in helping, please let me know here. Thanks!
