### 2024-02-12 #5549 @Kevin-Mattheus-Moerman
@sfmig thanks for the quick reply Sofía! You can call `@editorialbot generate my checklist` to get started. Let me know if you need help. Thanks!
