### 2024-02-20 FabianPlum/OmniTrax#8 @sfmig
[JOSS] Data sharing - broken links
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/5549#issue-1754074694) for this repo.

The repo provides a lot of very useful data, like sample footage and trained models, but I found that some of the links to the data were broken (specifically in `docs/example_footage.md` and `docs/tutorial-tracking.md`).

I used [this precommit hook](https://github.com/tcort/markdown-link-check) as a quick way to check all links in markdown files, below the summary output:

I leave you the full output below for reference (it may be useful to pin down the specific links).

  <summary>Full output</summary>
  Markdown Link Check......................................................Failed
- hook id: markdown-link-check
- exit code: 1

(node:47806) [DEP0040] DeprecationWarning: The `punycode` module is deprecated. Please use a userland alternative instead.
(Use `node --trace-deprecation ...` to show where the warning was created)

FILE: docs/CUDA_installation_guide.md
  [✓] https://en.wikipedia.org/wiki/CUDA
  [✓] https://github.com/AlexeyAB/darknet
  [✓] https://github.com/DeepLabCut/DeepLabCut-live
  [✓] https://developer.nvidia.com/cuda-11.2.0-download-archive?target_os=Windows&target_arch=x86_64&target_version=10&target_type=exelocal
  [✓] https://developer.nvidia.com/rdp/cudnn-archive
  [✓] ../images/omnitrax_logo.svg#gh-dark-mode-only
  [✓] ../images/omnitrax_logo_light.svg#gh-light-mode-only
  [✓] CUDA_installation_images/CUDA_01.PNG
  [✓] CUDA_installation_images/CUDA_02.PNG
  [✓] CUDA_installation_images/CUDA_03.PNG
  [✓] CUDA_installation_images/CUDA_04.PNG
  [✓] CUDA_installation_images/CUDA_05.PNG
  [✓] CUDA_installation_images/CUDA_06.PNG
  [✓] CUDA_installation_images/CUDA_07.PNG
  [✓] CUDA_installation_images/CUDA_08.PNG
  [✓] CUDA_installation_images/CUDA_09.PNG
  [✓] CUDA_installation_images/CUDA_10.PNG
  [✓] CUDA_installation_images/CUDA_11.PNG
  [✓] CUDA_installation_images/CUDA_12.PNG
  [✓] CUDA_installation_images/CUDA_13.PNG
  [✓] CUDA_installation_images/CUDA_14.PNG

  21 links checked.

FILE: .github/CONTRIBUTING.md
  [✓] #introduction
  [✓] #getting-started
  [✓] #contributing
  [✓] #reporting-bugs
  [✓] #suggesting-enhancements
  [✓] #code-contributions
  [✓] #contact
  [✓] mailto:fabian.plum18@imperial.ac.uk
  [✓] https://github.com/FabianPlum/OmniTrax/tree/main/docs
  [✓] https://github.com/FabianPlum/OmniTrax/issues
  [✓] https://github.com/FabianPlum/OmniTrax/issues/new/choose
  [✖] https://github.com/omnitrax/omnitrax

  12 links checked.

  ERROR: 1 dead links found!
  [✖] https://github.com/omnitrax/omnitrax → Status: 404

FILE: docs/example_footage.md
  [✓] ../README.md
  [✓] https://drive.google.com/file/d/1I0vla-CyTYpNIKNRJIzegxJ44WGyQ291/view?usp=share_link
  [✓] https://drive.google.com/file/d/1f417gbG7nt3xMIfKZgmr-3gPEUm_DPoJ/view?usp=share_link
  [✓] https://drive.google.com/file/d/1pa4hD-64JroByLavQZCvigMs7RGVxyvs/view?usp=share_link
  [✓] https://drive.google.com/file/d/1n-SRw7hswtMpaaXoGLuu_i1SFPgCBKoh/view?usp=share_link
  [✓] https://drive.google.com/file/d/1esvN2C4Egto_kZFWg5qGsETVphaa3aSi/view?usp=share_link
  [✓] https://drive.google.com/file/d/1rn4WUGyh8gotdC_UuVHIhqlqVn6sOhob/view?usp=share_link
  [✖] https://drive.google.com/file/d/10d2YuEpx62UOU8oQ1179XVxuKZOCbUZ5/view?usp=share_link
  [✓] https://drive.google.com/file/d/1X5fNkaEkALo1lgAu4HsKgzyZSAEapIq_/view?usp=share_link
  [✓] https://drive.google.com/file/d/109u6MyJFlLaiHaf08OavPWS8I6KvmPee/view?usp=share_link
  [✓] https://drive.google.com/file/d/1izoE7bLScQODYloV5B6bwzWtJ4jcqp1K/view?usp=sharing
  [✓] https://drive.google.com/file/d/1XzZmgkBUKeA3Q1YeYGMbwQoqsYtYgcjF/view?usp=sharing
  [✓] https://drive.google.com/drive/folders/14wBXFhV1KI4nD_TZXTrZssXqdOsWuDwk?usp=sharing
  [✓] ../images/omnitrax_logo.svg#gh-dark-mode-only
  [✓] ../images/omnitrax_logo_light.svg#gh-light-mode-only
  [✓] ../images/preview_tracking.gif
  [✓] ../images/multi_ants_online_tracking_&_pose_estimation.gif

  17 links checked.

  ERROR: 1 dead links found!
  [✖] https://drive.google.com/file/d/10d2YuEpx62UOU8oQ1179XVxuKZOCbUZ5/view?usp=share_link → Status: 404

FILE: .github/ISSUE_TEMPLATE/bug_report.md
  No hyperlinks found!

  0 links checked.
(node:47807) [DEP0040] DeprecationWarning: The `punycode` module is deprecated. Please use a userland alternative instead.
(Use `node --trace-deprecation ...` to show where the warning was created)

FILE: docs/tutorial-tracking.md
  [✓] https://github.com/DeepLabCut/DeepLabCut-live
  [✓] https://www.mackenziemathislab.org/dlc-modelzoo
  [✓] trained_networks.md
  [✓] example_footage.md
  [✓] https://github.com/AlexeyAB/darknet
  [✓] https://github.com/DeepLabCut/DeepLabCut
  [✓] ../README.md
  [✓] ../images/example_ant_recording.mp4
  [✓] https://drive.google.com/drive/folders/1PSseMeClcYIe9dcYG-JaOD2CzYceiWdl?usp=sharing
  [✓] https://www.blender.org/download/lts/3-3/
  [✓] CUDA_installation_guide.md
  [✓] https://github.com/FabianPlum/FARTS
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.2.1
  [✓] tutorial-pose-estimation.md
  [✓] https://en.wikipedia.org/wiki/Kalman_filter
  [✓] https://en.wikipedia.org/wiki/Hungarian_algorithm
  [✓] https://github.com/FabianPlum/blenderMotionExport
  [✓] https://github.com/Amudtogal
  [✓] ../example_scripts/Tracking_Dataset_Processing.ipynb
  [✖] ..images/example_ant_recording.mp4
  [✓] ../example_scripts/example_ant_recording
  [✓] https://choosealicense.com/licenses/mit/
  [✓] ../images/omnitrax_logo.svg#gh-dark-mode-only
  [✓] ../images/omnitrax_logo_light.svg#gh-light-mode-only
  [✓] ../images/use_01.jpg
  [✓] ../images/use_02.jpg
  [✓] ../images/use_03.jpg
  [✓] ../images/masking_01.png
  [✓] ../images/masking_02.png
  [✓] ../images/masking_03.png
  [✓] ../images/use_04.gif
  [✓] ../images/example_ant_tracked.gif
  [✖] ../example_scripts/_heatmap_of_ground_truth_tracks.svg
  [✓] ../images/ase_01.jpg
  [✓] ../images/ase_new_02.jpg

  35 links checked.

  ERROR: 2 dead links found!
  [✖] ..images/example_ant_recording.mp4 → Status: 400
  [✖] ../example_scripts/_heatmap_of_ground_truth_tracks.svg → Status: 400

FILE: docs/trained_networks.md
  [✓] https://github.com/AlexeyAB/darknet
  [✓] https://drive.google.com/drive/folders/1PSseMeClcYIe9dcYG-JaOD2CzYceiWdl?usp=sharing
  [✓] https://drive.google.com/drive/folders/11QXseJwISdodSnXJV6fwM97XfT2aXx2y?usp=sharing
  [✓] https://drive.google.com/drive/folders/1wQcfLlDUvnWthyzbvyVy9oqyTZ2F-JFo?usp=sharing
  [✓] https://drive.google.com/drive/folders/1U9jzOpjCcu6wDfTEH3uQqGKPxW_QzHGz?usp=sharing
  [✓] https://drive.google.com/drive/folders/1eXAowtyBsqGEjvmQE1YlSeHJ6AGBwpUs?usp=share_link
  [✓] https://github.com/AlexeyAB/darknet/wiki/YOLOv4-model-zoo
  [✓] https://github.com/DeepLabCut/DeepLabCut
  [✓] https://drive.google.com/drive/folders/1or1TF3tvi1iIzldEAia3G2RNKY5J7Qz4?usp=sharing
  [✓] https://drive.google.com/drive/folders/1FY3lAkAisOG_RIUBuaynz1OjBkzjH5LL?usp=sharing
  [✓] https://drive.google.com/file/d/1IH9R9PgJMYteigsrMi-bZnz4IMcydtWU/view?usp=sharing
  [✓] https://drive.google.com/drive/folders/1-DHkegHiTkWbO7YboXxDC5tU4Aa71-9z?usp=share_link
  [✓] https://drive.google.com/drive/folders/1BLulUYkwww7SfzXgSSVM71GLI4dQysP5?usp=share_link
  [✓] https://arxiv.org/abs/1605.03170
  [✓] https://www.mackenziemathislab.org/dlc-modelzoo
  [✓] tutorial-pose-estimation.md
  [✓] https://choosealicense.com/licenses/mit/
  [✓] ../images/omnitrax_logo.svg#gh-dark-mode-only
  [✓] ../images/omnitrax_logo_light.svg#gh-light-mode-only

  19 links checked.

FILE: .github/ISSUE_TEMPLATE/feature_request.md
  No hyperlinks found!

  0 links checked.

FILE: README.md
  [✓] https://github.com/FabianPlum/OmniTrax/releases
  [✓] https://github.com/FabianPlum/OmniTrax
  [✓] https://www.python.org/
  [✓] https://app.travis-ci.com/github/FabianPlum/OmniTrax
  [✓] https://github.com/FabianPlum/FARTS
  [✓] https://youtu.be/YXxM4QRaCDU
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.3.1
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.3.0
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.2.3
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.2.2
  [✓] https://github.com/FabianPlum/OmniTrax/blob/main/docs/tutorial-tracking.md
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.2.1
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.2.0
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.1.3
  [✓] https://www.blender.org/download/lts/3-3/
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.1.2
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.1.1
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.1
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.0.2
  [✓] https://github.com/DeepLabCut/DeepLabCut-live
  [✓] https://github.com/FabianPlum/OmniTrax/releases/tag/V_0.0.1
  [✓] https://download.blender.org/release/Blender2.92/
  [✓] https://developer.nvidia.com/cuda-11.2.0-download-archive?target_os=Windows&target_arch=x86_64&target_version=10&target_type=exelocal
  [✓] https://developer.nvidia.com/rdp/cudnn-archive
  [✓] https://www.tensorflow.org/install/source#gpu
  [✓] https://www.blender.org/download/release/Blender3.3/blender-3.3.1-windows-x64.msi/
  [✓] docs/CUDA_installation_guide.md
  [✓] https://github.com/FabianPlum/OmniTrax/releases/download/V_0.2.3/omni_trax.zip
  [✓] docs/tutorial-tracking.md
  [✓] docs/tutorial-pose-estimation.md
  [✓] https://github.com/AlexeyAB/darknet
  [✓] https://drive.google.com/drive/folders/1PSseMeClcYIe9dcYG-JaOD2CzYceiWdl?usp=sharing
  [✓] images/example_ant_recording.mp4
  [✓] docs/trained_networks.md
  [✓] docs/example_footage.md
  [✓] https://choosealicense.com/licenses/mit/
  [✓] https://app.travis-ci.com/FabianPlum/OmniTrax.svg?branch=main
  [✓] images/omnitrax_logo.svg#gh-dark-mode-only
  [✓] images/omnitrax_logo_light.svg#gh-light-mode-only
  [✓] images/preview_tracking.gif
  [✓] images/single_ant_1080p_POSE_track_0.gif
  [✓] images/single_ant_1080p_POSE_track_0_skeleton.gif
  [✓] images/omnitrax_demo_screen_updated.jpg
  [✓] images/install_01.jpg
  [✓] images/install_02.jpg
  [✓] images/install_03.jpg
  [✓] images/install_04.jpg
  [✓] images/install_05.jpg
  [✓] images/install_06.jpg
  [✓] images/use_01.jpg
  [✓] images/use_02.jpg
  [✓] images/use_03.jpg
  [✓] images/use_04.gif

  56 links checked.
(node:47808) [DEP0040] DeprecationWarning: The `punycode` module is deprecated. Please use a userland alternative instead.
(Use `node --trace-deprecation ...` to show where the warning was created)

FILE: docs/tutorial-pose-estimation.md
  [✓] https://github.com/DeepLabCut/DeepLabCut-live
  [✓] https://www.mackenziemathislab.org/dlc-modelzoo
  [✓] trained_networks.md
  [✓] example_footage.md
  [✓] https://github.com/AlexeyAB/darknet
  [✓] https://github.com/DeepLabCut/DeepLabCut
  [✓] ../README.md
  [✓] https://github.com/FabianPlum/FARTS
  [✓] https://drive.google.com/file/d/156t8r3ZHrkzC72jZapFl9OBFPqNIvIXg/view?usp=share_link
  [✓] https://drive.google.com/drive/folders/1-DHkegHiTkWbO7YboXxDC5tU4Aa71-9z?usp=share_link
  [✓] https://drive.google.com/file/d/1izoE7bLScQODYloV5B6bwzWtJ4jcqp1K/view?usp=sharing
  [✓] https://drive.google.com/drive/folders/1PSseMeClcYIe9dcYG-JaOD2CzYceiWdl?usp=sharing
  [✓] https://drive.google.com/drive/folders/1FY3lAkAisOG_RIUBuaynz1OjBkzjH5LL?usp=sharing
  [✓] tutorial-tracking.md
  [✓] https://choosealicense.com/licenses/mit/
  [✓] ../images/omnitrax_logo.svg#gh-dark-mode-only
  [✓] ../images/omnitrax_logo_light.svg#gh-light-mode-only
  [✓] ../images/single_ant_1080p_POSE_track_0.gif
  [✓] ../images/single_ant_1080p_POSE_track_0_skeleton.gif
  [✓] ../images/VID_20220201_160304_50%25_POSE_fullframe.gif
  [✓] ../images/multi_ants_online_tracking_&_pose_estimation.gif
  [✓] ../images/Human_Tracking.gif
  [✓] ../images/Human_POSE_fullframe.gif

  23 links checked.

FILE: paper/paper.md
  [✓] ../images/omnitrax_demo_screen.jpg

  1 links checked.
