### 2024-02-21 c5f12f7
added option to pass tracks in user-defined order
