+++ docs/tutorial-tracking.md
@@ -252,10 +252,10 @@ annotation approaches.
 ![](../images/example_ant_tracked.gif)
 
 _tracking output of **OmniTrax** for the first 1000 frames
-of [example_ant_recording.mp4](..images/example_ant_recording.mp4), using
+of [example_ant_recording.mp4](../images/example_ant_recording.mp4), using
 the [single class ant detector (trained on synthetic data)](https://drive.google.com/drive/folders/1PSseMeClcYIe9dcYG-JaOD2CzYceiWdl?usp=sharing)_
 
-![](../example_scripts/_heatmap_of_ground_truth_tracks.svg)
+![](../example_scripts/example_ant_recording_heatmap_of_ground_truth_tracks.svg)
 
 _generated heatmap from the [tracked outputs](../example_scripts/example_ant_recording)_ of
 [example_ant_recording.mp4](..images/example_ant_recording.mp4)
