### 2024-02-21 FabianPlum/OmniTrax#8 @FabianPlum
Just to quicky add to this - thanks so much @sfmig for pointing me towards the link checking tool. I now included an action with the repo to auto-check all links. Very useful advice!
