### 2024-02-21 FabianPlum/OmniTrax#8 @sfmig
glad you liked it! 🤩

You may want to check [this cookiecutter template](https://github.com/UCL-ARC/python-tooling), which includes commonly used Github action workflows and precommit hooks for Python packages. We use it often at the software development group at UCL. Funnily enough it does not include the link checker right now but probs it's a good addition 😅

You may know this but just in case, I leave you more info on [cookiecutter](https://libraries.io/pypi/cookiecutter) templates and on [precommit hooks](https://pre-commit.com/). The former is a CLI tool that lets you create projects from a project template. The latter ones basically execute code (typically linting or formating) right before you commit. This can be very helpful when developing collaboratively and they are very easy to set up.
