### 2024-02-21 FabianPlum/OmniTrax#8 @sfmig
Just FYI (and to continue the 🐰 🕳️ ): I have been pointed to [this link checker](https://github.com/lycheeverse/lychee-action) which seems nicer and a bit more supported than the one I sent
