### 2024-02-22 12db349
Updated links to latest version for installation guide | closes #10
