### 2024-02-22 816e38e
Added note on terminal use in Ubuntu | closes #13
