### 2024-02-22 61ad103
Added lookup table for operating system support and moved updates to the bottom of the page | closes #14
