### 2024-02-22 0db7eb2
Removed redundant info and updated formatting | closes #16
