### 2024-02-24 f57378f
Updated broken links and fixed file permission issue | resolves #17 , resolves #18
