+++ .github/workflows/links.yml
@@ -1,7 +1,6 @@
 name: Links checker
 
 on:
-  push:
   repository_dispatch:
   workflow_dispatch:
   schedule:
