### 2024-02-28 eebbd5b
updated required Blender version number to avoid incorrectly flagging mismatch | closes #19
