### 2024-02-29 2cb68e2
Added additional discussion on state of the field and tracking approach | #20
