+++ paper/paper.bib
@@ -71,3 +71,41 @@
     volume = {14},
     year = {2023}
     }
+
+@article{Kalafatic2001,
+author = {Kalafatic, Zoran and Ribaric, Slobodan and Stanisavljevic, Vladimir},
+doi = {10.1109/ICIAP.2001.957031},
+file = {:C\:/Users/PlumStation/AppData/Roaming/Blender Foundation/Blender/3.3/scripts/addons/omni_trax/sources/A_system_for_tracking_laboratory_animals_based_on_optical_flow_and_active_contours.pdf:pdf},
+isbn = {076951183X},
+journal = {Proceedings - 11th International Conference on Image Analysis and Processing, ICIAP 2001},
+pages = {334--339},
+title = {{A system for tracking laboratory animals based on optical flow and active contours}},
+year = {2001}
+}
+
+@article{Walter2021,
+author = {Walter, Tristan and Couzin, Iain D.},
+doi = {10.7554/eLife.64000},
+file = {:C\:/Users/PlumStation/AppData/Roaming/Blender Foundation/Blender/3.3/scripts/addons/omni_trax/sources/elife-64000-v2 (1).pdf:pdf},
+issn = {2050084X},
+journal = {eLife},
+pages = {1--73},
+pmid = {33634789},
+title = {{Trex, a fast multi-animal tracking system with markerless identi cation, and 2d estimation of posture and visual elds}},
+volume = {10},
+year = {2021}
+}
+
+@article{Perez-Escudero2014,
+author = {P{\'{e}}rez-Escudero, Alfonso and Vicente-Page, Juli{\'{a}}n and Hinz, Robert C and Arganda, Sara and {De Polavieja}, Gonzalo G},
+doi = {10.1038/nmeth.2994},
+file = {:C\:/Users/PlumStation/AppData/Roaming/Blender Foundation/Blender/3.3/scripts/addons/omni_trax/sources/nmeth.2994.pdf:pdf},
+issn = {15487105},
+journal = {Nature Methods},
+number = {7},
+pages = {743--748},
+title = {{IdTracker: Tracking individuals in a group by automatic identification of unmarked animals}},
+volume = {11},
+year = {2014}
+}
+
