### 2024-02-29 0821afb
Updated contribution guidelines and added option to submit trained models | closes #22
