### 2024-02-29 6d74734
Updated testrunner and documentation for local testing | resolves #24
