### 2024-02-29 32b825f
Updated readme by merging introduction text from CONTRIBUTION file | #22
