### 2024-02-29 167841e
Fixed global variable assignment for pose export | #27
