### 2024-03-01 3a9d156
Began restructuring codebase | #26
