+++ tracking/tracker.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 try:
-    from omni_trax.kalman_filter_new import KalmanFilter
+    from omni_trax.tracking.kalman_filter_new import KalmanFilter
 except ModuleNotFoundError:
     from kalman_filter_new import KalmanFilter
 from scipy.optimize import linear_sum_assignment
