+++ setup_state.txt
@@ -1,1 +1,0 @@
-setup_complete=False
