### 2024-03-01 51992be
Restructured complete project for readability | #26
