+++ __init__.py
@@ -1,31 +1,12 @@
-import bpy
-from bpy.props import BoolProperty as BoolProperty
-from bpy.props import StringProperty as StringProperty
-from bpy.props import IntProperty as IntProperty
-from bpy.props import FloatProperty as FloatProperty
-from bpy.props import EnumProperty as EnumProperty
-
 # check for installed packages and if they are missing, install them now
 # after the initial installation, dependency checks will be disabled
 # in setup_state.txt set setup_completed to "False" to trigger a system check / reinstall when enabling omni_trax
-from omni_trax.utils import check_packages
-from omni_trax.utils.CUDA_checks import check_CUDA_installation
+from .utils.setup import check_packages
+from .utils.setup.CUDA_checks import check_CUDA_installation
 
-import numpy as np
-import cv2
-import os
-import time
-import yaml
-import tensorflow as tf
-import platform
-
-from omni_trax.utils.omni_trax_utils import scale_detections, convertBack, cvDrawBoxes, setInferenceDevive
-
-# kalman imports
-from omni_trax.tracking.tracker import Tracker
-
-# YOLO handler
-from omni_trax.utils.YOLOader import YOLOader
+from .utils import compute
+from .utils import track
+from .utils import pose
 
 bl_info = {
     "name": "omni_trax",
@@ -34,1751 +15,21 @@ bl_info = {
     "blender": (2, 92, 0),
     "version": (0, 3, 1),
     "location": "",
-    "warning": "RUN IN ADMINISTRATOR MODE DURING INSTALLATION!",
+    "author": "Fabian Plum",
+    "warning": "RUN IN ADMINISTRATOR MODE DURING INSTALLATION! "
+               "Additional dependencies may conflict with other custom Add-ons.",
     "category": "motion capture"
 }
 
 
-class OMNITRAX_PT_ComputePanel(bpy.types.Panel):
-    """
-    Select (CUDA) computation device to run inference
-    """
-    bl_label = "Computational Device"
-    bl_space_type = "CLIP_EDITOR"
-    bl_region_type = "TOOLS"
-    bl_category = "OmniTrax"
-
-    try:
-        physical_devices = tf.config.list_physical_devices()
-        print("INFO: Found computational devices:\n", physical_devices)
-
-        devices = []
-        for d, device in enumerate(physical_devices):
-            if device.device_type == "GPU":
-                devices.append(("GPU_" + str(d - 1), "GPU_" + str(d - 1), "Use GPU for inference (requires CUDA)"))
-            else:
-                if bpy.app.version_string == "2.92.0":
-                    devices.append(("CPU_" + str(d), "CPU", "Use CPU for inference"))
-    except:
-        print("WARNING: Did not find suitable inference devices.")
-
-    bpy.types.Scene.compute_device = EnumProperty(
-        name="",
-        description="Select inference device",
-        items=devices,
-        default=devices[-1][0])
-
-    def draw(self, context):
-        layout = self.layout
-
-        col = layout.column(align=True)
-        col.label(text="Select computational device")
-        col.separator()
-        row = col.row(align=True)
-        row.prop(context.scene, "compute_device")
-
-
-class OMNITRAX_OT_DetectionOperator(bpy.types.Operator):
-    """
-    Run the (yolo) detection based buffer-and-recover tracking pipeline
-    RESTART Track: Begin tracking from defined Start-Frame, overwrite prior tracks if present
-    TRACK: Resume tracking from previous state
-    """
-    bl_idname = "scene.detection_run"
-    bl_label = "Run Detection"
-
-    restart_tracking: BoolProperty(
-        name="Restart Tracking",
-        description="Re-initialises tracker with new settings. WARNING: Current identities will NOT be retained!",
-        default=False)
-
[…]
