+++ utils/track/__init__.py
@@ -1,0 +1,11 @@
+from . import ui, operators
+
+
+def register():
+    ui.register()
+    operators.register()
+
+
+def unregister():
+    ui.unregister()
+    operators.unregister()
