+++ tests/utils.test.py
@@ -2,7 +2,7 @@ import sys
 import unittest
 
 try:
-    from omni_trax.omni_trax_utils import *
+    from omni_trax.utils.omni_trax_utils import *
 
 
     class TestUtils(unittest.TestCase):
@@ -30,5 +30,6 @@ try:
     if success.errors or success.failures:
         raise Exception
 
-except Exception:
+except Exception as e:
+    print(e)
     sys.exit(1)
