+++ tests/tracker.test.py
@@ -2,7 +2,7 @@ import sys
 import unittest
 
 try:
-    from omni_trax.tracker import *
+    from omni_trax.utils.track.tracker import *
 
 
     class TestTracker(unittest.TestCase):
@@ -125,5 +125,6 @@ try:
     if success.errors or success.failures:
         raise Exception
 
-except Exception:
+except Exception as e:
+    print(e)
     sys.exit(1)
