### 2024-03-01 7ccb76e
Added info for local execution
