### 2024-03-01 #5549 @sfmig
Hi @Kevin-Mattheus-Moerman, @FabianPlum,
I just finished going through the items in the checklist - the test item is still unchecked because tests are not passing locally for me yet, but I think Fabian is working on refactoring the code. Happy to cross that one out once it's done :)

Thanks for tagging me, it's been fun reviewing this cool tool 🤩
