### 2024-03-01 de57c94
Added exception handling for alternative imports | #26 closes #31
