+++ tests/tracker.test.py
@@ -1,11 +1,17 @@
 import sys
 import unittest
 import numpy as np
+import os
 
 try:
     try:
         from omni_trax.utils.track.tracker import Tracker
     except:
+        from inspect import getsourcefile
+        from os.path import abspath
+
+        project_dir = os.path.dirname(os.path.dirname(getsourcefile(lambda: 0)))
+        sys.path.append(project_dir)
         from utils.track.tracker import Tracker
 
 
