### 2024-03-02 7ed2b72
Added pyproject file and unified formatting
