+++ utils/track/operators.py
@@ -19,13 +19,15 @@ class OMNITRAX_OT_DetectionOperator(bpy.types.Operator):
     RESTART Track: Begin tracking from defined Start-Frame, overwrite prior tracks if present
     TRACK: Resume tracking from previous state
     """
+
     bl_idname = "scene.detection_run"
     bl_label = "Run Detection"
 
     restart_tracking: BoolProperty(
         name="Restart Tracking",
         description="Re-initialises tracker with new settings. WARNING: Current identities will NOT be retained!",
-        default=False)
+        default=False,
+    )
 
     def execute(self, context):
         # start by checking all necessary paths are provided before attempting to do anything
@@ -33,19 +35,32 @@ class OMNITRAX_OT_DetectionOperator(bpy.types.Operator):
         try:
             video = bpy.path.abspath(clip.filepath)
         except:
-            self.report({'ERROR'}, 'Open a video to track from your drive, then click TRACK / RESTART Tracking again')
-            return {'CANCELLED'}
+            self.report(
+                {"ERROR"},
+                "Open a video to track from your drive, then click TRACK / RESTART Tracking again",
+            )
+            return {"CANCELLED"}
 
         # check required YOLO paths
         yolo_cfg_path_temp = bpy.path.abspath(context.scene.detection_config_path)
-        if not os.path.isfile(yolo_cfg_path_temp) or yolo_cfg_path_temp.split(".")[-1] != "cfg":
-            self.report({'ERROR'}, 'Provide the .cfg file of a suitable YOLOv3 or V4 model')
-            return {'CANCELLED'}
+        if (
+            not os.path.isfile(yolo_cfg_path_temp)
+            or yolo_cfg_path_temp.split(".")[-1] != "cfg"
+        ):
+            self.report(
+                {"ERROR"}, "Provide the .cfg file of a suitable YOLOv3 or V4 model"
+            )
+            return {"CANCELLED"}
 
         yolo_weights_path_temp = bpy.path.abspath(context.scene.detection_weights_path)
-        if not os.path.isfile(yolo_weights_path_temp) or yolo_weights_path_temp.split(".")[-1] != "weights":
-            self.report({'ERROR'}, 'Provide the .weights file of a trained YOLOv3 or V4 model')
-            return {'CANCELLED'}
+        if (
+            not os.path.isfile(yolo_weights_path_temp)
+            or yolo_weights_path_temp.split(".")[-1] != "weights"
+        ):
+            self.report(
+                {"ERROR"}, "Provide the .weights file of a trained YOLOv3 or V4 model"
+            )
+            return {"CANCELLED"}
 
         """
         Check which compute device is selected and set it as active
@@ -55,13 +70,19 @@ class OMNITRAX_OT_DetectionOperator(bpy.types.Operator):
         # load darknet with compiled DLLs for windows for either GPU or CPU inference from respective path
         if context.scene.compute_device.split("_")[0] == "GPU":
             required_CUDA_version = "11.2"
-            CUDA_match = check_CUDA_installation(required_CUDA_version=required_CUDA_version)
+            CUDA_match = check_CUDA_installation(
+                required_CUDA_version=required_CUDA_version
+            )
 
             if not CUDA_match:
-                self.report({'ERROR'}, 'No matching CUDA version found! Refer to console for full error message')
-                return {'CANCELLED'}
+                self.report(
+                    {"ERROR"},
+                    "No matching CUDA version found! Refer to console for full error message",
+                )
+                return {"CANCELLED"}
 
             from ..darknet import darknet as darknet
+
             darknet.set_compute_device(int(context.scene.compute_device.split("_")[1]))
         else:
             if platform.system() != "Linux":
[…]
