+++ utils/compute/ui.py
@@ -7,6 +7,7 @@ class OMNITRAX_PT_ComputePanel(bpy.types.Panel):
     """
     Select (CUDA) computation device to run inference
     """
+
     bl_label = "Computational Device"
     bl_space_type = "CLIP_EDITOR"
     bl_region_type = "TOOLS"
@@ -19,7 +20,13 @@ class OMNITRAX_PT_ComputePanel(bpy.types.Panel):
         devices = []
         for d, device in enumerate(physical_devices):
             if device.device_type == "GPU":
-                devices.append(("GPU_" + str(d - 1), "GPU_" + str(d - 1), "Use GPU for inference (requires CUDA)"))
+                devices.append(
+                    (
+                        "GPU_" + str(d - 1),
+                        "GPU_" + str(d - 1),
+                        "Use GPU for inference (requires CUDA)",
+                    )
+                )
             else:
                 if bpy.app.version_string == "2.92.0":
                     devices.append(("CPU_" + str(d), "CPU", "Use CPU for inference"))
@@ -30,7 +37,8 @@ class OMNITRAX_PT_ComputePanel(bpy.types.Panel):
         name="",
         description="Select inference device",
         items=devices,
-        default=devices[-1][0])
+        default=devices[-1][0],
+    )
 
     def draw(self, context):
         layout = self.layout
