+++ utils/track/ui.py
@@ -16,56 +16,67 @@ class OMNITRAX_PT_DetectionPanel(bpy.types.Panel):
         name="Config path",
         description="Path to your YOLO network config file",
         default="",  # bpy.data.filepath.rpartition("\\")[0],
-        subtype="FILE_PATH")
+        subtype="FILE_PATH",
+    )
     bpy.types.Scene.detection_weights_path = StringProperty(
         name="Weights path",
         description="Path to your YOLO network weights file",
         default="",  # bpy.data.filepath.rpartition("\\")[0],
-        subtype="FILE_PATH")
+        subtype="FILE_PATH",
+    )
     bpy.types.Scene.detection_data_path = StringProperty(
         name="Data path",
         description="Path to your YOLO network data file",
         default="",  # bpy.data.filepath.rpartition("\\")[0],
-        subtype="FILE_PATH")
+        subtype="FILE_PATH",
+    )
     bpy.types.Scene.detection_names_path = StringProperty(
         name="Names path",
         description="Path to your YOLO network names file",
         default="",  # bpy.data.filepath.rpartition("\\")[0],
-        subtype="FILE_PATH")
+        subtype="FILE_PATH",
+    )
     bpy.types.Scene.detection_enforce_constant_size = BoolProperty(
         name="constant detection sizes",
         description="Check to enforce constant detection sizes. This DOES NOT affect the actual inference, only the resulting regions of interest.",
-        default=True)
+        default=True,
+    )
     bpy.types.Scene.detection_constant_size = IntProperty(
         name="constant detection sizes (px)",
         description="Constant detection size in pixels. All detections will be rescaled to this value in both width and height.",
-        default=60)
+        default=60,
+    )
     bpy.types.Scene.detection_min_size = IntProperty(
         name="minimum detection sizes (px)",
         description="If the width or height of a detection is below this threshold, it will be discarded. This can be useful to decrease noise. Keep the value at 0 if this is not needed.",
-        default=15)
+        default=15,
+    )
     bpy.types.Scene.detection_network_width = FloatProperty(
         name="Network width",
         description="YOLO network input width. Larger network inputs allow for smaller detected subjects but will increase inference time and memory usage. [MUST BE MULTIPLE OF 32]",
         default=480,
         step=3200,  # step size given in 1/100
         min=32,
-        precision=0)  # set to 0 emulate integer property (which currently does not support steps)
+        precision=0,
+    )  # set to 0 emulate integer property (which currently does not support steps)
     bpy.types.Scene.detection_network_height = FloatProperty(
         name="Network height",
         description="YOLO network input height. Larger network inputs allow for smaller detected subjects but will increase inference time and memory usage. [MUST BE MULTIPLE OF 32]",
         default=480,
         step=3200,  # step size given in 1/100
         min=32,
-        precision=0)  # set to 0 emulate integer property (which currently does not support steps)
+        precision=0,
+    )  # set to 0 emulate integer property (which currently does not support steps)
     bpy.types.Scene.detection_activation_threshold = FloatProperty(
         name="Confidence threshold",
         description="Detection confidence threshold. A higher confidence leads to fewer false positives but more missed detections.",
-        default=0.5)
+        default=0.5,
+    )
     bpy.types.Scene.detection_nms = FloatProperty(
         name="Non-maximum suppression",
[…]

[…]
