### 2024-03-02 78fc37c
Added pyproject file and unified formatting | #29, closes #30
