### 2024-03-02 4a7b59a
Updated remaining docstrings | closes #29
