+++ utils/darknet_sub_process/opencv_direct_darknet.py
@@ -4,15 +4,15 @@ import numpy as np
 import time
 
 img = cv.imread(
-    "C:/Users/Legos/AppData/Roaming/Blender Foundation/Blender/3.2/scripts/addons/omni_trax/darknet_sub_process/ant3.jpg"
+	"C:/Users/Legos/AppData/Roaming/Blender Foundation/Blender/3.2/scripts/addons/omni_trax/darknet_sub_process/ant3.jpg"
 )
 cv.imshow("window", img)
 cv.waitKey(1)
 
 # Give the configuration and weight files for the model and load the network.
 net = cv.dnn.readNetFromDarknet(
-    "C:/Users/Legos/Documents/PhD/Blender/OmniTrax/trained_networks/atta_single_class/yolov4-big_and_small_ants_480.cfg",
-    "C:/Users/Legos/Documents/PhD/Blender/OmniTrax/trained_networks/atta_single_class/yolov4-big_and_small_ants_HPC_final.weights",
+	"C:/Users/Legos/Documents/PhD/Blender/OmniTrax/trained_networks/atta_single_class/yolov4-big_and_small_ants_480.cfg",
+	"C:/Users/Legos/Documents/PhD/Blender/OmniTrax/trained_networks/atta_single_class/yolov4-big_and_small_ants_HPC_final.weights",
 )
 net.setPreferableBackend(cv.dnn.DNN_BACKEND_CUDA)
 net.setPreferableTarget(cv.dnn.DNN_TARGET_CUDA)
@@ -21,12 +21,12 @@ net.setPreferableTarget(cv.dnn.DNN_TARGET_CUDA)
 
 # Load names of classes and get random colors
 classes = (
-    open(
-        "C:/Users/Legos/Documents/PhD/Blender/OmniTrax/trained_networks/atta_single_class/obj.names"
-    )
-    .read()
-    .strip()
-    .split("\n")
+	open(
+		"C:/Users/Legos/Documents/PhD/Blender/OmniTrax/trained_networks/atta_single_class/obj.names"
+	)
+	.read()
+	.strip()
+	.split("\n")
 )
 np.random.seed(42)
 colors = np.random.randint(0, 255, size=(len(classes), 3), dtype="uint8")
@@ -52,7 +52,7 @@ print("time=", t - t0)
 
 print(len(outputs))
 for out in outputs:
-    print(out.shape)
+	print(out.shape)
 
 boxes = []
 confidences = []
@@ -60,29 +60,29 @@ classIDs = []
 h, w = img.shape[:2]
 
 for output in outputs:
-    for detection in output:
-        scores = detection[5:]
-        classID = np.argmax(scores)
-        confidence = scores[classID]
-        if confidence > 0.5:
-            box = detection[:4] * np.array([w, h, w, h])
-            (centerX, centerY, width, height) = box.astype("int")
-            x = int(centerX - (width / 2))
-            y = int(centerY - (height / 2))
-            box = [x, y, int(width), int(height)]
-            boxes.append(box)
-            confidences.append(float(confidence))
-            classIDs.append(classID)
+	for detection in output:
+		scores = detection[5:]
+		classID = np.argmax(scores)
+		confidence = scores[classID]
+		if confidence > 0.5:
+			box = detection[:4] * np.array([w, h, w, h])
+			(centerX, centerY, width, height) = box.astype("int")
+			x = int(centerX - (width / 2))
+			y = int(centerY - (height / 2))
+			box = [x, y, int(width), int(height)]
+			boxes.append(box)
+			confidences.append(float(confidence))
+			classIDs.append(classID)
 
 indices = cv.dnn.NMSBoxes(boxes, confidences, 0.2, 0.4)
[…]
