+++ utils/darknet_sub_process/darknet_evaluation_post_inference.py
@@ -10,242 +10,242 @@ from multiprocessing import Pool
 
 
 def compare_points(gt, detection, max_dist=25):
-	match = False
-	px_distance = distance.euclidean(gt, detection)
-	if px_distance <= max_dist:
-		match = True
-	return match, px_distance
+    match = False
+    px_distance = distance.euclidean(gt, detection)
+    if px_distance <= max_dist:
+        match = True
+    return match, px_distance
 
 
 def compare_frame(
-	frame_gt, frame_detections, max_dist=0.05, network_shape=[None, None], confidence=0
+    frame_gt, frame_detections, max_dist=0.05, network_shape=[None, None], confidence=0
 ):
-	# strip away all sub threshold detections!
-	frame_detections = [f for f in frame_detections if f[1] > confidence]
+    # strip away all sub threshold detections!
+    frame_detections = [f for f in frame_detections if f[1] > confidence]
 
-	matches_gt = np.ones(len(frame_gt))
-	matches_det = np.ones(len(frame_detections))
-	below_thresh = 0
-	detection_distances = []
+    matches_gt = np.ones(len(frame_gt))
+    matches_det = np.ones(len(frame_detections))
+    below_thresh = 0
+    detection_distances = []
 
-	# now strip all empty entries from the ground truth
+    # now strip all empty entries from the ground truth
 
-	for i in range(len(matches_gt)):
-		min_dist = max_dist
-		for j in range(len(matches_det)):
-			if network_shape[0] is not None:
-				norm_frame_detection = [
-					frame_detections[j][2][0] / network_shape[0],
-					frame_detections[j][2][1] / network_shape[1],
-				]
+    for i in range(len(matches_gt)):
+        min_dist = max_dist
+        for j in range(len(matches_det)):
+            if network_shape[0] is not None:
+                norm_frame_detection = [
+                    frame_detections[j][2][0] / network_shape[0],
+                    frame_detections[j][2][1] / network_shape[1],
+                ]
 
-			else:
-				norm_frame_detection = frame_detections[j][2][0:2]
+            else:
+                norm_frame_detection = frame_detections[j][2][0:2]
 
-			match, px_dist = compare_points(
-				gt=frame_gt[i][0:2], detection=norm_frame_detection, max_dist=max_dist
-			)
+            match, px_dist = compare_points(
+                gt=frame_gt[i][0:2], detection=norm_frame_detection, max_dist=max_dist
+            )
 
-			if match:
-				matches_gt[i] = 0
-				matches_det[j] = 0
-				if px_dist < min_dist:
-					min_dist = px_dist
+            if match:
+                matches_gt[i] = 0
+                matches_det[j] = 0
+                if px_dist < min_dist:
+                    min_dist = px_dist
 
-		if min_dist < max_dist:
-			detection_distances.append(min_dist)
+        if min_dist < max_dist:
+            detection_distances.append(min_dist)
 
-	missed_detections = int(np.sum(matches_gt))
-	false_positives = int(np.sum(matches_det)) - below_thresh
+    missed_detections = int(np.sum(matches_gt))
+    false_positives = int(np.sum(matches_det)) - below_thresh
 
-	if len(detection_distances) == 0:
-		mean_detection_distance = 0
-	else:
-		mean_detection_distance = np.mean(np.array(detection_distances))
+    if len(detection_distances) == 0:
+        mean_detection_distance = 0
+    else:
+        mean_detection_distance = np.mean(np.array(detection_distances))
 
-	return len(frame_gt), missed_detections, false_positives, mean_detection_distance
+    return len(frame_gt), missed_detections, false_positives, mean_detection_distance
 
 
 def getThreads():
[…]
