### 2024-03-02 963f3b4
Costant detection size parameter only active when enabled | #28
