### 2024-03-04 #5549 @FabianPlum
Hi @sfmig and @rizarae-p! Thanks so much for your detailed feedback and many useful suggestions. I think the repo has improved a lot because of your input!
I just tested the local unit testing again under Windows for Blender 2.92, 3.3.0, and 3.3.16 and they now all ran through as expected with the updated exception handling after some imports had failed earlier.

Let me know if your last concerns are addressed with these updates!
All the best
Fabi
