### 2024-03-04 2411bd6
Added exception handling for low or missing parsed fps values.
