+++ paper/paper.bib
@@ -32,6 +32,16 @@
     bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
+@article{Kuhn1955,
+author = {Kuhn, Harold W},
+journal = {Naval Research Logistics (NRL)},
+keywords = {hungarian},
+mendeley-tags = {hungarian},
+title = {{The Hungarian method for the assignment problem}},
+volume = {52},
+year = {1955}
+}
+
 @article{Mathisetal2018,
     title = {DeepLabCut: markerless pose estimation of user-defined body parts with deep learning},
     author = {Alexander Mathis and Pranav Mamidanna and Kevin M. Cury and Taiga Abe and Venkatesh N. Murthy and Mackenzie W. Mathis and Matthias Bethge},
