### 2024-03-13 #5549 @Kevin-Mattheus-Moerman
@FabianPlum have a look a the above post-review action points. Could you work on those? Let me know if you have any questions.

In addition, can you check the below:
- [ ] Check if this [10.1002/nav.3800020109](https://doi.org/10.1002/nav.3800020109) is a suitable DOI for: `The Hungarian method for the assignment problem`
- [ ] You can likely use this one [10.48550/arXiv.1804.02767](https://doi.org/10.48550/arXiv.1804.02767) for the arXiv paper.
