### 2024-03-14 #5549 @FabianPlum
Fixed the two references! Thanks a lot.
I don't think there is a reasonable DOI for Blender.

> * [x]  Check if this [10.1002/nav.3800020109](https://doi.org/10.1002/nav.3800020109) is a suitable DOI for: `The Hungarian method for the assignment problem`
> * [x]  You can likely use this one [10.48550/arXiv.1804.02767](https://doi.org/10.48550/arXiv.1804.02767) for the arXiv paper.
