+++ paper/paper.bib
@@ -27,7 +27,7 @@
     url = {http://arxiv.org/abs/1804.02767},
     eprinttype = {arXiv},
     eprint = {1804.02767},
-    doi = {10.48550/arXiv.1804.02767}
+    doi = {10.48550/arXiv.1804.02767},
     timestamp = {Mon, 13 Aug 2018 16:48:24 +0200},
     biburl = {https://dblp.org/rec/journals/corr/abs-1804-02767.bib},
     bibsource = {dblp computer science bibliography, https://dblp.org}
@@ -39,7 +39,7 @@ journal = {Naval Research Logistics (NRL)},
 keywords = {hungarian},
 mendeley-tags = {hungarian},
 title = {{The Hungarian method for the assignment problem}},
-doi = {10.1002/nav.3800020109}
+doi = {10.1002/nav.3800020109},
 volume = {52},
 year = {1955}
 }
