### 2024-03-14 #5549 @FabianPlum
Hi @Kevin-Mattheus-Moerman!
All tasks done. The latest release with all additions included (1.0.0) is available via [GitHub](https://github.com/FabianPlum/OmniTrax/releases/tag/V_1.0.0), as well as [10.5281/zenodo.10817891](https://zenodo.org/doi/10.5281/zenodo.10817891)

> ## Additional Author Tasks After Review is Complete
> * [x] Double check authors and affiliations (including ORCIDs)
> * [x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [x] Archive the release on Zenodo/figshare/etc and post the DOI here.
> * [x] Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [x] Make sure that the license listed for the archive is the same as the software license.
